task 3.1.5 is completed

+++ src/main/java/ru/kata/spring/boot_security/demo/controllers/AdminController.java
@@ -1,15 +1,17 @@
 package ru.kata.spring.boot_security.demo.controllers;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.Authentication;
-import org.springframework.stereotype.Controller;
-import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
+import ru.kata.spring.boot_security.demo.exceptions.NoSuchUserException;
+import ru.kata.spring.boot_security.demo.models.Role;
 import ru.kata.spring.boot_security.demo.models.User;
 import ru.kata.spring.boot_security.demo.services.RoleService;
 import ru.kata.spring.boot_security.demo.services.UserService;
 
-@Controller
+import java.util.List;
+
+@RestController
+@RequestMapping("/api")
 public class AdminController {
 
     private final UserService userService;
@@ -21,33 +23,44 @@ public class AdminController {
         this.roleService = roleService;
     }
 
-    @GetMapping("/admin")
-    public String showAllUsers(Model model, Authentication authentication) {
-        User user = (User) authentication.getPrincipal();
-        model.addAttribute("authUser", user);
-        model.addAttribute("newUser", new User());
-        model.addAttribute("allUsers", userService.listUsers());
-        model.addAttribute("allRoles", roleService.listRoles());
-        model.addAttribute("activeTab", "usersTable");
-        return "admin";
+    @GetMapping("/users")
+    public List<User> showAllUsers() {
+        return userService.listUsers();
     }
 
-    @PostMapping("/admin")
-    public String addUser(@ModelAttribute("newUser") User user, Model model) {
-        model.addAttribute("activeTab", "newUser");
+    @GetMapping("/users/{id}")
+    public User showUser(@PathVariable long id) {
+        User user = userService.findById(id);
+        if (user == null) {
+            throw new NoSuchUserException("There's no user with ID = " + id + " in the database");
+        }
+        return user;
+    }
+
+    @PostMapping("/users")
+    public User addUser(@RequestBody User user) {
         userService.add(user);
-        return "redirect:/admin";
+        return user;
     }
 
-    @PatchMapping("/admin")
-    public String updateUser(@ModelAttribute("user") User user) {
+    @PutMapping("/users")
+    public User updateUser(@RequestBody User user) {
         userService.update(user);
-        return "redirect:/admin";
+        return user;
     }
 
-    @DeleteMapping("/admin")
-    public String deleteUser(@RequestParam("id") Long id) {
+    @DeleteMapping("/users/{id}")
+    public String deleteUser(@PathVariable long id) {
+        User user = userService.findById(id);
+        if (user == null) {
+            throw new NoSuchUserException("There's no user with ID = " + id + " in the database");
+        }
         userService.delete(id);
-        return "redirect:/admin";
+        return "User with ID = " + id + " was deleted";
+    }
+
+    @GetMapping("/roles")
+    public List<Role> getAllRoles() {
+        return roleService.listRoles();
     }
 }
